Add option to encode traveltime as RGBA png values.
The travel time in seconds is encoded in the color channels of the
generated png.  Disregarding the alpha channel, we encode the value
in base 255, using each channel for one digit.

Alpha channel is always set to 255 to avoid browser optimizations that

This commit also adds html and javascript for an in-process experiment
with animating travelsheds with the output of the service.