Accept capitalized params for `wms` endpoint
- Leaflet (or someone) seems keen on passing capitalized
parameters to the GT site. These are redirected to the
transit API, but Transit used to barf on capitals. Previously
the params were made small by the GT site code, but when
redirecting instead via nginx (for docker-compose), this
doesn't happen. The Transit API now accepts capitalized
params for the `travelshed/wms` endpoint.